// frontend/src/app/models/fault.model.ts
export interface Fault {
  id: string;
  title: string;
  reporterName: string;
  severity: string;
  status: string;
  responsiblePerson: string;
  description: string;
  resolution: string;
  resolvedAt: string;
  department: string;
  systemName: string;
  alertStatus: string;
  alertLogId: string;
  createdAt: string;
  updatedAt: string;
}